requirements.txt
httpx~=0.28.1
ujson~=5.10.0
aiogram~=3.16.0
python-dotenv~=1.0.1
fastapi~=0.115.6
starlette~=0.41.3
python-jose~=3.3.0
SQLAlchemy~=2.0.36
passlib~=1.7.4
pydantic~=2.10.4
uvicorn~=0.34.0
psycopg2-binary~=2.9.10
datetime~=5.5
python-multipart~=0.0.20